Fix incorrect comparison of integer array instance values
This part of the test was performed incorrectly, the for loop uses j, and the value of i fixes the comparison to `short` only and ignores the other types; This caused the test to pass by chance.
This has been rewritten with an outer loop that goes over the different type names (name) and j that iterates over the array indices.

The values of `alpha` were also incorrect and did not reflect the multiplier used in the test dictionary.

Another point to note is in the case of signed char, the pointer `m_schar_array2` is implicitly converted to `std::basic_string` while the fixed size array `m_schar_array` is not. Using the getter `get_schar_array2` however works here and we only compare using the instance retrieved this way and skip comparing `c.m_schar_array2` for this reason.